refactor: consolidate interfaces and types into a single file
